% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_cox_pv}
\alias{compute_cox_pv}
\title{Calculate p-values associated with survival analyses}
\usage{
compute_cox_pv(d, e, idx_c, idx_n, thresh_func = ector::m3sd)
}
\arguments{
\item{d}{a transcriptomic dataset}

\item{e}{a clinical dataset}

\item{idx_c}{a vector of ID of cancerous patients}

\item{idx_n}{a vector of ID of normal patients}

\item{thresh_func}{the function used to determine the threshold}
}
\description{
This function calculates, for each gene, the Cox and logrank p-values and the hazard ratio from survival analyses
}
