% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_cox_pv_grp}
\alias{compute_cox_pv_grp}
\title{Calculate p-values associated with survival analysis of grouped patients}
\usage{
compute_cox_pv_grp(e, idx_c, x)
}
\arguments{
\item{e}{a clinical dataset}

\item{idx_c}{a vector of ID of cancerous patients}

\item{x}{a vector of number of genes ectopically expressed in patients}
}
\description{
This function calculates the Cox, logrank p-values and hazard ratio from the survival analysis of patients depending on the number of genes ectopically expressed
}
