% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ector_clinicals}
\alias{ector_clinicals}
\title{Clinicals data frame}
\format{A data frame with XXX rows and 7 columns}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
ector_clinicals
}
\description{
The `ector_clinicals` data frame describes clinical information.
}
\keyword{datasets}
