% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ector_tissue_specific_genes}
\alias{ector_tissue_specific_genes}
\title{Tissue specific genes data frame}
\format{A data frame with 3433 rows and 7 columns}
\source{
\url{http://epimed.univ-grenoble-alpes.fr/database}
}
\usage{
ector_tissue_specific_genes
}
\description{
The `ector_tissue_specific_genes` data frame describes tissue-specific genes. 
The rows correspond to genes while the columns correspond to their related informations, 
including position in GRCh38 reference genome.
The data presentation is inspired by BED (Browser Extensible Data) format.
Tissue specific genes are determined using threshold algorithme [ref. EpiMed]. ons.
}
\keyword{datasets}
