% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ector_transcriptome}
\alias{ector_transcriptome}
\title{Transcriptome matrix}
\format{A matrix with 3000 rows and 44 columns}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
ector_transcriptome
}
\description{
The ector_transcriptome matrix...
}
\keyword{datasets}
