% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mnsd}
\alias{mnsd}
\title{Compute mean plus n sd on a vector}
\usage{
mnsd(x, n)
}
\arguments{
\item{x}{a vector of gene expression}

\item{n}{a numeric corresponding to the number of sd to add}
}
\description{
This function computes mean plus n sd on a vector
}
