% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_gene_expr}
\alias{plot_gene_expr}
\title{Plot gene expression of normal and tumoral samples}
\usage{
plot_gene_expr(gs, d, e, idx_n, thresh_func = ector::m3sd)
}
\arguments{
\item{gs}{a gene symbol}

\item{d}{a transcriptomic dataset}

\item{e}{a clinical dataset}

\item{idx_n}{a vector of ID of normal patients}

\item{thresh_func}{the function used to determine the threshold}
}
\description{
This function draws a plot in which each sample is represented by a dot
}
