% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival curves}
\usage{
plot_survival(gs, d, e, idx_c, idx_n, thresh_func = ector::m3sd)
}
\arguments{
\item{gs}{a gene symbol}

\item{d}{a transcriptomic dataset}

\item{e}{a clinical dataset}

\item{idx_c}{a vector of ID of cancerous patients}

\item{idx_n}{a vector of ID of normal patients}

\item{thresh_func}{the function used to determine the threshold}
}
\description{
This function draws the Kaplan-Meier curves of patients in which a given gene is activated or not
}
