% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_survival_grp}
\alias{plot_survival_grp}
\title{Plot survival curves of grouped patients}
\usage{
plot_survival_grp(e, idx_c, x)
}
\arguments{
\item{e}{a clinical dataset}

\item{idx_c}{a vector of ID of cancerous patients}

\item{x}{a vector of patients groups (numeric)}
}
\description{
This function draws the Kaplan-Meier curves of patients grouped according to the number of genes ectopically expressed
}
