% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prop_ectopic_expr}
\alias{prop_ectopic_expr}
\title{Proportion of patients with ectopic gene expression}
\usage{
prop_ectopic_expr(d, idx_c, idx_n, thresh_func = ector::m3sd)
}
\arguments{
\item{d}{a transcriptomic dataset}

\item{idx_c}{a vector of ID of cancerous patients}

\item{idx_n}{a vector of ID of normal patients}

\item{thresh_func}{the function used to determine the threshold}
}
\description{
This function calculates, for each gene, the proportion of patients in which the expression is ectopic
}
